requirements.txt
streamlit>=1.32.0
google-genai>=0.3.0
openai>=1.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
duckduckgo-search>=5.0.0
requests>=2.31.0
chromadb>=0.4.22
pysqlite3-binary>=0.5.2; sys_platform == 'linux'
typing-extensions>=4.9.0
